requirements.txt
# CloneMe Requirements - Exact Tested Versions
# Tested and verified on Python 3.11.6
# These exact versions ensure maximum compatibility and stability

# Core AI and LangChain packages
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.52.2
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
curl_cffi==0.11.2
dataclasses-json==0.6.7
discord-protos==0.0.2
distro==1.9.0
duckduckgo_search==8.0.2
filetype==1.2.0
frozenlist==1.6.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
googlesearch-python==1.3.0
greenlet==3.2.3
groq==0.26.0
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-anthropic==0.3.15
langchain-community==0.3.24
langchain-core==0.3.64
langchain-google-genai==2.0.10
langchain-groq==0.3.2
langchain-ollama==0.3.3
langchain-openai==0.3.19
langchain-text-splitters==0.3.8
langsmith==0.3.45
lxml==5.4.0
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
numpy==2.2.6
ollama==0.5.1
openai==1.84.0
orjson==3.10.18
packaging==24.2
primp==0.15.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyNaCl==1.5.0
pyparsing==3.2.3
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
watchdog==6.0.0
yarl==1.20.0
zstandard==0.23.0
# Matrix Platform Support
matrix-nio>=0.24.0

# Discord selfbot library - specific commit for stability
# Note: This will be installed separately by setup.py
# git+https://github.com/dolfies/discord.py-self.git@71609f4f62649d18bdf14f0e286b7e62bc605390
